🐛 fix(worker-profiling-app): restore two-server setup, fix deflate worker
Revert to separate dev (port 8081) + proxy (port 8082) commands.

proxy-server.ts:
- Builds and serves /datadog-worker.js via webpack-dev-middleware so
  the deflate worker initialises correctly (fixes the 30s timeout that
  aborted RUM startup when WORKER_STRING was empty)
- Handles CORS so the page on 8081 can POST/GET to 8082
- Parses profile multipart payloads inline (CJS-free, ESM busboy)
- Broadcasts profile summaries over SSE GET /events

webpack.config.cjs (renamed from .js):
- Back to minimal config: no busboy/SSE, no second compiler
- Renamed to .cjs so Node treats it as CommonJS with package type:module

package.json:
- Add type:module (silences Node ESM detection warning for proxy-server.ts)
- Remove webpack-dev-middleware (only needed by proxy, uses root modules)
- proxy script: node proxy-server.ts

main.ts: workerUrl + proxy + EventSource all point to PROXY_ORIGIN (8082)
