+++ test/apps/worker-profiling/package.json
@@ -1,9 +1,10 @@
 {
   "name": "worker-profiling-app",
   "private": true,
+  "type": "module",
   "scripts": {
-    "build": "webpack --config ./webpack.config.js",
-    "dev": "webpack serve --config ./webpack.config.js",
+    "build": "webpack --config ./webpack.config.cjs",
+    "dev": "webpack serve --config ./webpack.config.cjs",
     "proxy": "node proxy-server.ts"
   },
   "peerDependencies": {
@@ -22,7 +23,6 @@
     "typescript": "6.0.2",
     "webpack": "5.107.2",
     "webpack-cli": "6.0.1",
-    "webpack-dev-middleware": "7.4.2",
     "webpack-dev-server": "5.2.1"
   },
   "volta": {
