+++ test/apps/worker-profiling/src/index.html
@@ -405,7 +405,7 @@
       <div class="info-box">
         <strong>Requires:</strong> Chromium Canary with<br>
         <code>--enable-features=DocumentPolicyInDedicatedWorker,<br>ProfilerAPIForDedicatedWorker</code><br><br>
-        <strong>Proxy:</strong> built into the dev server — just <code>yarn dev</code> is enough.
+        <strong>Proxy server:</strong> run <code>yarn proxy</code> in a second terminal.
       </div>
     </div>
